+++ src/plugins/dam/components/uploader/delegate.js
@@ -173,15 +173,21 @@ class Delegate extends AbstractDelegate {
   }
 
   handleCreditApplyAll(credit, files, initialValues) {
+    const config = { schemas, assetCount: Object.keys(files).length };
     const data = {
       fields: ['credit'],
       values: { credit },
     };
-    this.dispatch(initialize(this.getFormName(), { ...initialValues, credit }, 'credit'));
-    this.dispatch(patchAssets(data, files, this.config));
+    this.dispatch(initialize(
+      this.getFormName(),
+      { ...initialValues, credit: { value: credit } },
+      'credit',
+    ));
+    this.dispatch(patchAssets(data, files, config));
   }
 
   handleExpiresAtApplyAll(expiresAt, files, initialValues) {
+    const config = { schemas, assetCount: Object.keys(files).length };
     const data = {
       fields: ['expires_at'],
       values: { expires_at: expiresAt },
@@ -190,7 +196,7 @@ class Delegate extends AbstractDelegate {
       this.getFormName(),
       { ...initialValues, expiresAt },
     ));
-    this.dispatch(patchAssets(data, files, this.config));
+    this.dispatch(patchAssets(data, files, config));
   }
 
   getFormName() {