+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 
 ## vN.N.N
+* Youtube Block: Choosing start time throws error 
 * Clear image for Audio & Article block not functioning on Safari.