+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 
 ## vN.N.N
-* Youtube Block: Choosing start time throws error 
+* Youtube Block: Choosing start time throws error
 * Clear image for Audio & Article block not functioning on Safari.
+* CMS - Uploading assets is throwing an error for some users
 
 
 ## v0.13.1